MVP - Segunda version/requirements.txt
streamlit==1.52.2
pandas==2.3.3
databricks-sql-connector==4.2.4
python-dotenv==1.2.1
plotly==6.5.1
numpy==2.4.1
matplotlib==3.10.8
scikit-learn==1.8.0